+++ graphgenBP.h
@@ -5,11 +5,11 @@
 
 typedef struct /* the bipartite graph data structure */
 {
-	long n; // numver of vertices in both sides
-	long nrows; // number of vertices in the left side
-	long m; // number of edges
-	long* vtx_pointer; // an array of size n+1 storing the pointer in endV array
-    long* endV; //an array of size m that stores the second vertex of an edge.
+	int n; // numver of vertices in both sides
+	int nrows; // number of vertices in the left side
+	int m; // number of edges
+	int* vtx_pointer; // an array of size n+1 storing the pointer in endV array
+    int* endV; //an array of size m that stores the second vertex of an edge.
 	double* weight; // not used in unweighted graph
 } graph;
